pyproject.toml
[project]
name = "quant-research-os"
version = "0.1.0"
description = "Codex skill generator for stage review workflows"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
